nct6775: Add support for multiple virtual temperature sources
For virtual temperatures, the actual temperature values are written
by software, presumably by the BIOS. This functionality is (as of
right now) supported on NCT6791D, NCT6792D, and NCT6793D. On those chips,
the temperatures are written into registers 0xea..0xef on page 0.
This is known to be used on some Asus motherboards. Report the 'virtual'
temperatures for all monotoring sources to address this situation.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>